requirements.txt
Automat==20.2.0
Babel==2.11.0
Bottleneck==1.3.7
Brotli==1.0.9
Django==5.1.2
GitPython==3.1.37
HeapDict==1.0.1
Markdown==3.4.1
MarkupSafe==2.1.3
Protego==0.1.16
PuLP==2.9.0
PyDispatcher==2.0.5
PyJWT==2.8.0
PyNaCl==1.5.0
PyQt5==5.15.10
PyQt5-sip==12.13.0
PyQtWebEngine==5.15.6
PySocks==1.7.1
PyYAML==6.0.1
Pygments==2.15.1
QDarkStyle==3.2.3
QtAwesome==1.2.2
QtPy==2.4.1
Rtree==1.0.1
SQLAlchemy==2.0.30
Scrapy==2.11.1
Send2Trash==1.8.2
Unidecode==1.2.0
XlsxWriter==3.2.0
aext-assistant==4.0.15
aext-assistant-server==4.0.15
aext-core==4.0.15
aext-core-server==4.0.15
aext-panels==4.0.15
aext-panels-server==4.0.15
aext-share-notebook==4.0.15
aext-share-notebook-server==4.0.15
aext-shared==4.0.15
aiobotocore==2.12.3
aiohttp==3.9.5
aioitertools==0.7.1
aiosignal==1.2.0
alabaster==0.7.16
altair==5.0.1
altgraph==0.17.4
anaconda-anon-usage==0.4.4
anaconda-catalogs==0.2.0
anaconda-client==1.12.3
anaconda-cloud-auth==0.5.1
anaconda-navigator==2.6.0
anaconda-project==0.11.1
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
archspec==0.2.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.8.1
astroid==2.14.2
astropy==6.1.0
astropy-iers-data==0.2024.6.3.0.31.14
asttokens==2.0.5
async-lru==2.0.4
atomicwrites==1.4.0
attrs==23.1.0
autopep8==2.0.4
bcrypt==3.2.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
black==24.4.2
bleach==4.1.0
blinker==1.6.2
bokeh==3.4.1
boltons==23.0.0
botocore==1.34.69
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
chardet==4.0.0
charset-normalizer==2.0.4
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
colorcet==3.1.0
comm==0.2.1
conda==24.7.1
conda-build==24.5.1
conda-content-trust==0.2.0
conda-index==0.5.0
conda-libmamba-solver==24.1.0
conda-pack==0.7.1
conda-package-handling==2.3.0
conda-package-streaming==0.10.0
conda-repo-cli==1.0.88
conda-token==0.5.0+1.g2209e04
constantly==23.10.4
contourpy==1.2.0
cookiecutter==2.6.0
cryptography==42.0.5
cssselect==1.2.0
cycler==0.11.0
cytoolz==0.12.2
dask==2024.5.0
dask-expr==1.1.0
datashader==0.16.2
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.8
distributed==2024.5.0
distro==1.9.0
docstring-to-markdown==0.11
docutils==0.18.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fastjsonschema==2.16.2
filelock==3.13.1
flake8==7.0.0
flask==3.0.3
fonttools==4.51.0
frozendict==2.4.2
frozenlist==1.4.0
fsspec==2024.3.1
gensim==4.3.2
gitdb==4.0.7
google-auth==2.38.0
google-auth-oauthlib==1.2.1
greenlet==3.0.1
gspread==6.1.4
h11==0.14.0
h5py==3.11.0
holoviews==1.19.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
hvplot==0.10.0
hyperlink==21.0.0
idna==3.7
imagecodecs==2023.1.23
imageio==2.33.1
imageio-ffmpeg==0.5.1
imagesize==1.4.1
imbalanced-learn==0.12.3
importlib-metadata==7.0.1
incremental==22.10.0
inflection==0.5.1
iniconfig==1.1.1
intake==0.7.0
intervaltree==3.1.0
ipykernel==6.28.0
ipython==8.25.0
ipython-genutils==0.2.0
ipywidgets==7.8.1
isort==5.13.2
itemadapter==0.3.0
itemloaders==1.1.0
itsdangerous==2.2.0
jaraco.classes==3.2.1
jedi==0.18.1
jellyfish==1.0.1
jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
json5==0.9.6
jsonpatch==1.33
jsonpointer==2.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-client==8.6.0
jupyter-console==6.6.3
jupyter-core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.0
jupyter-server==2.14.1
jupyter-server-terminals==0.4.4
jupyterlab==4.0.11
jupyterlab-pygments==0.1.2
jupyterlab-server==2.25.1
jupyterlab-widgets==1.0.0
keyring==24.3.1
kiwisolver==1.4.4
lazy-loader==0.4
lazy-object-proxy==1.10.0
lckr-jupyterlab-variableinspector==3.1.0
libarchive-c==2.9
libmambapy==1.5.8
linkify-it-py==2.0.0
llvmlite==0.42.0
lmdb==1.4.1
locket==1.0.0
lxml==5.2.1
lz4==4.3.2
markdown-it-py==2.2.0
matplotlib==3.8.4
matplotlib-inline==0.1.6
mccabe==0.7.0
mdit-py-plugins==0.3.0
mdurl==0.1.0
menuinst==2.1.1
mistune==2.0.4
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
more-itertools==10.1.0
mpmath==1.3.0
msgpack==1.0.3
multidict==6.0.4
multipledispatch==0.6.0
mutagen==1.47.0
mypy==1.10.0
mypy-extensions==1.0.0
navigator-updater==0.5.1
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
notebook==7.0.8
notebook-shim==0.2.3
numba==0.59.1
numexpr==2.8.7
numpy==1.26.4
numpydoc==1.7.0
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.2
overrides==7.4.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.0
panel==1.4.4
param==2.1.0
paramiko==2.8.1
parsel==1.8.1
parso==0.8.3
partd==1.4.1
pathspec==0.10.3
patsy==0.5.6
pdfminer.six==20231228
pdfplumber==0.11.4
pefile==2023.2.7
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.3.0
pip==24.0
pkce==1.0.3
pkginfo==1.10.0
platformdirs==3.10.0
plotly==5.22.0
pluggy==1.0.0
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyOpenSSL==24.0.0
py-cpuinfo==9.0.0
pyarrow==14.0.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.11.1
pycosat==0.6.6
pycparser==2.21
pycryptodomex==3.20.0
pyct==0.5.0
pycurl==7.45.2
pydantic==2.5.3
pydantic-core==2.14.6
pydeck==0.8.0
pydocstyle==6.3.0
pyerfa==2.0.1.4
pyflakes==3.2.0
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pylint==2.16.2
pylint-venv==3.0.3
pyls-spyder==0.4.0
pyngrok==7.2.3
pyodbc==5.0.1
pyparsing==3.0.9
pypdf2==3.0.1
pypdfium2==4.30.0
pytest==7.4.4
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==0.21.0
python-json-logger==2.0.7
python-lsp-black==2.0.0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.10.0
python-pptx==1.0.2
python-slugify==5.0.2
python-snappy==0.6.1
pytoolconfig==1.2.6
pytube==15.0.0
pytz==2024.1
pyviz-comms==3.0.2
pywavelets==1.5.0
pywin32==305.1
pywin32-ctypes==0.2.2
pywinpty==2.0.10
pyzmq==25.1.2
qstylizer==0.2.2
qtconsole==5.5.1
queuelib==1.6.2
referencing==0.30.2
regex==2023.10.3
requests==2.32.2
requests-file==1.5.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.5
rope==1.12.0
rpds-py==0.10.6
rsa==4.9
ruamel.yaml==0.17.21
ruamel-yaml-conda==0.17.21
s3fs==2024.3.1
scikit-image==0.23.2
scikit-learn==1.4.2
scipy==1.13.1
seaborn==0.13.2
semver==3.0.2
service-identity==18.1.0
setuptools==69.5.1
sip==6.7.12
six==1.16.0
smart-open==5.2.1
smmap==4.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
sphinx==7.3.7
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.10
spyder==5.5.1
spyder-kernels==2.5.0
sqlparse==0.5.1
stack-data==0.2.0
statsmodels==0.14.2
streamlit==1.32.0
sympy==1.13.1
tables==3.9.2
tabula-py==2.9.3
tabulate==0.9.0
tblib==1.7.0
tenacity==8.2.2
terminado==0.17.1
text-unidecode==1.3
textdistance==4.2.1
threadpoolctl==2.2.0
three-merge==0.1.1
tifffile==2023.4.12
tinycss2==1.2.1
tldextract==3.2.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
toolz==0.12.0
torch==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
truststore==0.8.0
twisted==23.10.0
twisted-iocpsupport==1.0.2
typing-extensions==4.11.0
tzdata==2023.3
uc-micro-py==1.0.1
ujson==5.10.0
unicodedata2==15.1.0
urllib3==2.2.2
w3lib==2.1.2
watchdog==4.0.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
werkzeug==3.0.3
whatthepatch==1.0.2
wheel==0.43.0
widgetsnbextension==3.6.6
win-inet-pton==1.1.0
wrapt==1.14.1
xarray==2023.6.0
xgboost==2.1.3
xlwings==0.31.4
xyzservices==2022.9.0
yapf==0.40.2
yarl==1.9.3
yt-dlp==2024.8.6
zict==3.0.0
zipp==3.17.0
zope.interface==5.4.0
zstandard==0.22.0
-ympy==1.13.1
